+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Upcoming release
 * Fix `sendPassword` & `presentScannable` invocations
 * Fix user preference parsing
+* Add `accessToken` to the account details
 
 ## 2.2.7
 * Updated to Android SDK 2.2.3