+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix user preference parsing
 * Add `accessToken` to the account details
 * Fix `TimeOfDayRange` parsing
+* Fix `updateDoNotDisturb` invocation
 
 ## 2.2.7
 * Updated to Android SDK 2.2.3